+++ boilerplate/scripts/webpack.config.babel.js
@@ -87,7 +87,7 @@ export default {
   ),
 
   plugins: [
-    new ExtractTextPlugin("[name].css", { disable: config.dev }),
+    new ExtractTextPlugin("[name].[hash].css", { disable: config.dev }),
     new webpack.DefinePlugin({ "process.env": {
       NODE_ENV: JSON.stringify(
         config.production ? "production" : process.env.NODE_ENV