+++ boilerplate/scripts/webpack.config.client.js
@@ -44,7 +44,7 @@ export default {
   output: {
     ...webpackConfig.output,
     libraryTarget: "var",
-    filename: "[name].js",
+    filename: "[name].[hash].js",
   },
   entry: {
     "statinamic-client": path.join(__dirname, "index-client"),